// frontend/src/api/generated/models/materialLedgerResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Manufacturing Information System API
 * API untuk aplikasi core MIS.
 * OpenAPI spec version: 1.0.0
 */

export interface MaterialLedgerResponse {
  id: string;
  material_id: string;
  transaction_type: string;
  /** @pattern ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$ */
  quantity: string;
  /** @pattern ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$ */
  unit_cost: string;
  reference_document: string;
  notes: string;
  created_at: string;
}